Performance: parallel preload, model rewrites, tuned thread/worker defaults
Three landing pieces from the perf-deep-dive sweep:

1. SQL model rewrites that don't change row counts, column names, or
   dtypes: QUALIFY-on-wide-row → narrow MAX(...) GROUP BY CTE, drop
   redundant window frames in favor of arithmetic derivation, ENUM/INT
   type alignment so joins go hash instead of BLOCKWISE_NL_JOIN, FILTER
   clause vs CASE pivot, min_by for first fielding position, ANY_VALUE
   for FD-determined columns. Affects 11 models.

2. scripts/preload_sources.py runs the 45 source parquet loads in
   parallel against bc.db before SQLMesh opens; init_db macro becomes
   a string of CREATE TABLE IF NOT EXISTS no-ops. Cuts init phase from
   ~63s to ~46s. Wired into scripts/perf_run.py as a pre-step.

3. Grid search over (BC_DUCKDB_THREADS, BC_CONCURRENT_TASKS) found
   t=14, w=6 (~6× core overcommit on a 14-core box) is the local
   optimum; bumped bc/config.py defaults from (7, 2) to (14, 6).
   Higher overcommit regresses. Knobs are env-driven so CI can pin
   alternate values without code changes.

End-to-end perf_run wall-clock: 4m06s → ~2m13s. Detail and full grid
results in notes/perf-deep-dive.md and notes/perf-profile-report.md.

+++ bc/models/intermediate/states/personnel_fielding_states.sql
@@ -6,7 +6,7 @@ MODEL (
     game_id VARCHAR,
     fielding_team_id TEAM_ID,
     fielding_side SIDE,
-    personnel_fielding_key BIGINT,
+    personnel_fielding_key INTEGER,
     start_event_id UTINYINT,
     end_event_id UINTEGER,
     player_id VARCHAR,
@@ -54,7 +54,10 @@ final AS (
         appearances.game_id,
         CASE WHEN appearances.side = 'Home' THEN games.home_team_id ELSE games.away_team_id END AS fielding_team_id,
         appearances.side AS fielding_side,
-        (games.game_key + ranges.start_event_id) * CASE WHEN appearances.side = 'Home' THEN 1 ELSE -1 END
+        -- INTEGER (not BIGINT) so downstream joins can hash on the raw column.
+        -- Range: ±(max(game_key) + 255). game_key is a row id over stg_games
+        -- (a few million max), well within INT range.
+        ((games.game_key + ranges.start_event_id) * CASE WHEN appearances.side = 'Home' THEN 1 ELSE -1 END)::INTEGER
         AS personnel_fielding_key,
         ranges.start_event_id,
         ranges.end_event_id,
